Refactor (#27)
* Develop (#25)

* feat/GURU : layout added for mobile with footer and toast added

* feat/GURU : layout added for mobile with footer and toast added

* feat/GURU : login validation with error message and show hide password done

* feat/ GURU: api bind to signup and login with validation and middlware done

* feat/GURU : middleware protection fully done

* feat/GURU : added logout and greeting

* feat/GURU: added mobile header

* Feature/guru api category (#22)

* feat: [GURU] on progress category

* feat: [GURU] category update

* feat: [GURU] delete category feature

* feat: [GURU] category pagination

* feat/GURU : small changes

* feat/GURU : mobile profile page aligned

* feat/GURU: wallet ui in progress

* feat/GURU : wallet design done

---------


* packge.lock.json

* added current tab on wallet for mock

* feat/GURU: added suspense

* fix/GURU: refactor the tab change

* fix/GURU: added the suspense

* fix/GURU: suspense layered

* fix/GURU: split the component in wallet

---------
